add litevent back in and set generic module references
